requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pymysql==1.1.0
pydantic-settings==2.2.1
httpx==0.27.0